+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [Unreleased]
+### Added
+- Add PHP 7.2 compatibility
+- Add `EnvelopeTag.width`
+- Add `EnvelopeTag.bankIdClaim`
+- Add `Envelope.sendCompleted`
+- Add `Envelope.timestampDocuments`
+- Add `EnvelopeTemplate.sendCompleted`
+- Add `EnvelopeTemplate.timestampDocuments`
+
 ## [1.0.2] - 2021-06-09
 ### Added
 - Add new `WebhookAttemptsEndpoint` and `WebhookAttempt` resource